# use signature-v2 for 'object perf' tests to avoid CPU using sha256